Chienup n

+++ mamilovebe/src/main/java/com/mamilove/response/dto/BillDashboardResponse.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 public class BillDashboardResponse {
 
     private String name;
-    private Double total;
+    private Long total;
 
 }
